CLI: Clean up the steps for the create-package
