+++ packages/frontity/src/commands/create-package.ts
@@ -1,14 +1,14 @@
 import chalk from "chalk";
 import { EventEmitter } from "events";
-import { Options } from "../steps/create-package/types";
 import {
-  ensurePackageDir,
+  Options,
   createPackageJson,
   createSrcIndexJs,
-  installPackage,
-  revertProgress,
-  isDirNotEmpty
-} from "../steps/create-package/steps";
+  installPackage
+} from "../steps/create-package";
+
+import { ensureProjectDir, revertProgress } from "../steps";
+import { join } from "path";
 
 export default async (options?: Options, emitter?: EventEmitter) => {
   // This functions will emit an event if an emitter is passed in options.
@@ -19,42 +19,37 @@ export default async (options?: Options, emitter?: EventEmitter) => {
   let step: Promise<any>;
   let dirExisted: boolean;
 
+  const { packagePath } = options;
+
   process.on("SIGINT", async () => {
-    if (typeof dirExisted !== "undefined") await revertProgress(options);
+    if (typeof dirExisted !== "undefined")
+      await revertProgress(dirExisted, packagePath);
   });
 
   try {
-    // 1. Make sure ./packages/[name] folder is empty.
-    step = isDirNotEmpty(options);
-    emit(
-      `Checking if ${chalk.yellow(options.packagePath)} is not empty.`,
-      step
-    );
-    const dirNotEmpty = await step;
-    if (dirNotEmpty)
-      throw new Error(`Folder "${options.packagePath}" must be empty.`);
-
-    // 2. Create ./packages/[name] folder.
-    step = ensurePackageDir(options);
-    emit(`Creating ${chalk.yellow(options.packagePath)} folder.`, step);
+    // 1. Create ./packages/[name] folder.
+    step = ensureProjectDir(join(packagePath, "src"));
+    emit(`Creating ${chalk.yellow(packagePath)} folder.`, step);
     dirExisted = await step;
 
-    // 3. Creates `package.json`.
+    // 2. Creates `package.json`.
     step = createPackageJson(options);
     emit(`Adding ${chalk.yellow("package.json")}.`, step);
     await step;
 
-    // 4. Creates `src/index.js`.
+    // 3. Creates `src/index.js`.
     step = createSrcIndexJs(options);
     emit(`Adding ${chalk.yellow("src/index.js")}.`, step);
     await step;
 
-    // 5. Install package
+    // 4. Install package
     step = installPackage(options);
     emit(`Installing package ${chalk.yellow(options.name)}.`, step);
     await step;
   } catch (error) {
-    if (typeof dirExisted !== "undefined") await revertProgress(options);
+    if (typeof dirExisted !== "undefined") {
+      await revertProgress(dirExisted, packagePath);
+    }
     if (emitter) emitter.emit("error", error);
     else throw error;
   }
