+++ packages/frontity/src/steps/index.ts
@@ -14,6 +14,7 @@ import {
 import { extract } from "tar";
 import fetch from "node-fetch";
 import { mergeRight } from "ramda";
+
 import { emitter } from "../utils/eventEmitter";
 import {
   isPackageNameValid,
